+++ Controllers/product.controllers.js
@@ -7,6 +7,10 @@ const Op=Sequelize.Op;
 
 exports.create=(req,res)=>{ 
 
+    if(!req.isAdmin){
+        return res.status(403).send({message:"OOPS! you are unauthorized to perform task"});
+    }
+
     const {name,description,cost,categoryId}=req.body;
     const product ={name,description,cost,categoryId};
 
@@ -98,6 +102,11 @@ exports.findOne=(req,res)=>{
 }
 
 exports.update =(req,res)=>{
+
+    if(!req.isAdmin){
+        return res.status(403).send({message:"OOPS! you are unauthorized to perform task"});
+    }
+
     const productId=req.params.id;
     const {name,description} =req.body;
     const product={};
@@ -120,6 +129,11 @@ exports.update =(req,res)=>{
 } 
 
 exports.delete= (req,res)=>{
+
+    if(!req.isAdmin){
+        return res.status(403).send({message:"OOPS! you are unauthorized to perform task"});
+    }
+
     const productId= req.params.id;
 
     Product.destroy({
